Hot code reloading
This is a simple socket app which listens to changes made to files you
specifiy like app.js or app.css for example. When you are coding and
saving your browser can auto refresh or hot reload. I set this up for
use in a PHP environment where browserify did not work out the box due
to the machine being on a proxy.